+++ packages/smb-commons/package.json
@@ -2,10 +2,9 @@
   "name": "peque-smb-commons",
   "version": "1.0.0",
   "description": "Peque.ts SMB Commons",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
   "scripts": {
-    "build": "pnpm clean && tsc",
     "clean": "rimraf .coverage dist",
     "test": "node --loader ts-node/esm ../../node_modules/uvu/bin.js src .test.ts",
     "lint": "eslint \"./{src,test}/**/*.ts\" --fix"
@@ -13,7 +12,8 @@
   "author": "Simone Di Cicco",
   "license": "ISC",
   "devDependencies": {
-    "peque-di": "workspace:*"
+    "peque-di": "workspace:*",
+    "peque-test": "workspace:*"
   },
   "engines": {
     "node": ">=16.0.0"
