+++ packages/smb-server/src/commands/subscribe.command.ts
@@ -1,5 +1,5 @@
 import { Injectable } from 'peque-di';
-import { EventService, ICommandService, ISubscribeCommand } from 'peque-smb-commons/src';
+import { EventService, ICommandService, ISubscribeCommand } from 'peque-smb-commons';
 
 import { SubscribeService } from '../services';
