+++ packages/smb-server/src/commands/welcome.command.ts
@@ -1,5 +1,5 @@
 import { Injectable } from 'peque-di';
-import { EventService, ICommandService, IWelcomeCommand } from 'peque-smb-commons/src';
+import { EventService, ICommandService, IWelcomeCommand } from 'peque-smb-commons';
 
 @Injectable()
 export class WelcomeCommand implements ICommandService {
