+++ packages/smb-server/src/di.ts
@@ -1,5 +1,5 @@
 import { DiContainer } from 'peque-di';
-import { Command, CommandEmitter, CommandParser, CommandReceiver, EventService } from 'peque-smb-commons/src';
+import { Command, CommandEmitter, CommandParser, CommandReceiver, EventService } from 'peque-smb-commons';
 
 import { MessageCommand, PublishCommand, SubscribeCommand, WelcomeCommand } from './commands';
 import { Broker } from './server/broker.class';
@@ -20,6 +20,7 @@ const providers = [
   PublishCommand,
   Broker,
 ];
+
 export const loadProviders = (): void => {
   for (const provider of providers) {
     DI.set(provider, provider.name);
