+++ packages/test/package.json
@@ -1,12 +1,11 @@
 {
-  "name": "peque-smb-server",
+  "name": "peque-test",
   "version": "1.0.0",
-  "description": "Peque.ts SMB Server",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "description": "Peque.ts Test utils",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
   "scripts": {
     "start": "ts-node src/index.ts",
-    "build": "pnpm clean && tsc",
     "clean": "rimraf .coverage dist",
     "test": "node --loader ts-node/esm ../../node_modules/uvu/bin.js src .test.ts",
     "lint": "eslint \"./{src,test}/**/*.ts\" --fix"
@@ -14,9 +13,7 @@
   "author": "Simone Di Cicco",
   "license": "ISC",
   "dependencies": {
-    "peque-di": "workspace:*",
-    "peque-smb-commons": "workspace:*",
-    "peque-test": "workspace:*"
+    "peque-di": "workspace:*"
   },
   "engines": {
     "node": ">=16.0.0"
